+++ lib/components/MiningDashboard.js
@@ -117,8 +117,8 @@ class MiningDashboard extends React.PureComponent {
           <Text type="span" className="col-lg-2">
             Miner: {mining ? 'on' : 'off'}
           </Text>
-          <Text type="span" className="col-lg-6">
-            Coinbase Address:
+          <Text type="span" className="col-lg-7">
+            Coinbase Address:{' '}
             {coinbaseAddress ? coinbaseAddress : 'No address set'}
           </Text>
         </div>
