+++ lib/constants.js
@@ -1,0 +1,2 @@
+export const PLUGIN_NAMESPACE = 'bp_simpleMining';
+export const UPDATE_COINBASE_ADDRESS = 'UPDATE_COINBASE_ADDRESS';
