+++ lib/index.js
@@ -1,10 +1,9 @@
 // Entry point for your plugin
 // This should expose your plugin's modules
-/* START IMPORTS */
 import MiningDashboard from './components/MiningDashboard';
-/* END IMPORTS */
-
-/* START EXPORTS */
+import { PLUGIN_NAMESPACE } from './constants';
+import { updateCoinbaseAddress } from './actions';
+import { minerReducer } from './reducers';
 
 export const metadata = {
   name: '@bpanel/simple-mining',
@@ -17,9 +16,6 @@ export const metadata = {
   sidebar: true
 };
 
-const PLUGIN_NAMESPACE = 'bp_simpleMining';
-const UPDATE_COINBASE_ADDRESS = 'UPDATE_COINBASE_ADDRESS';
-
 export const mapComponentState = {
   Panel: (state, map) =>
     Object.assign(map, {
@@ -31,8 +27,7 @@ export const mapComponentState = {
 export const mapComponentDispatch = {
   Panel: (dispatch, map) =>
     Object.assign(map, {
-      updateCoinbaseAddress: addr =>
-        dispatch({ type: UPDATE_COINBASE_ADDRESS, payload: addr })
+      updateCoinbaseAddress: addr => dispatch(updateCoinbaseAddress(addr))
     })
 };
 
@@ -46,30 +41,11 @@ export const getRouteProps = {
 };
 
 export const pluginReducers = {
-  [PLUGIN_NAMESPACE]: (state = { coinbaseAddress: '' }, action) => {
-    const { type, payload } = action;
-    let newState = { ...state };
-
-    switch (type) {
-      case UPDATE_COINBASE_ADDRESS: {
-        newState.coinbaseAddress = payload;
-        return newState;
-      }
-
-      default: {
-        return state;
-      }
-    }
-  }
+  [PLUGIN_NAMESPACE]: minerReducer
 };
 
 export const persistReducers = [PLUGIN_NAMESPACE];
 
-// a decorator for the Panel container component in our app
-// here we're extending the Panel's children by adding
-// our plugin's component (`MyComponent` below)
-// You'll want to make sure to import an actual component
-// This is what you need if you're making a new view/route
 export const decoratePanel = (Panel, { React, PropTypes }) => {
   return class extends React.Component {
     static displayName() {
